qppcpp library file cut into pieces to make compilation feasible
