+++ examples/geom/CMakeLists.txt
@@ -2,8 +2,12 @@ add_executable(vec vec.cpp)
 add_executable(geom1 geom1.cpp)
 add_executable(geom2 geom2.cpp)
 add_executable(geom6 geom6.cpp)
+target_link_libraries(geom6 qpp)
+
 add_executable(geom_oh geom_oh.cpp)
 add_executable(ngbr ngbr.cpp)
+target_link_libraries(ngbr qpp)
+
 #add_executable(si-ngbr si-ngbr.cpp)
 
 add_executable(shape shape.cpp)
