+++ modules/python/qppcpp.hpp
@@ -1,0 +1,13 @@
+#ifndef _QPP_QPPCPP_H
+#define _QPP_QPPCPP_H
+
+#include <boost/python.hpp>
+
+void qpp_export1();
+void qpp_export2();
+void qpp_export3();
+void qpp_export4();
+void qpp_export5();
+void qpp_export6();
+
+#endif
